server/requirements.txt
google-auth==2.19.1
google-oauth2-tool==0.0.3
aiohttp==3.8.4
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backports.pbkdf2==0.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bidict==0.22.1
capstone==5.0.0rc2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cli-helpers==2.2.1
click==8.1.3
colorama==0.4.4
colored-traceback==0.3.0
configobj==5.0.6
cryptography==37.0.4
delegator.py==0.1.1
distlib==0.3.6
filelock==3.12.2
Flask==2.2.2
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.2
frida==15.1.17
frida-tools==10.5.4
frozenlist==1.3.3
greenlet==2.0.1
h11==0.13.0
idna==3.3
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
litecli==1.9.0
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.4
objection==1.11.0
outcome==1.2.0
packaging==21.3
paramiko==2.12.0
pathlib2==2.3.7.post1
pexpect==4.8.0
platformdirs==3.5.3
plumbum==1.8.0
prompt-toolkit==3.0.28
psutil==5.9.4
ptyprocess==0.7.0
pwntools==4.8.0
pycparser==2.21
pycryptodome==3.12.0
pyelftools==0.29
Pygments==2.11.2
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.4.1
python-socketio==5.8.0
pywin32==305
redis==4.4.1
requests==2.27.1
ROPGadget==7.1
rpyc==5.2.3
selenium==4.3.0
semver==2.13.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.2
tabulate==0.8.10
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
unicorn==2.0.1
urllib3==1.26.8
virtualenv==20.23.0
wcwidth==0.2.5
webdriver-manager==3.8.5
websockets==11.0.2
Werkzeug==2.2.2
wsproto==1.1.0
yarl==1.9.2
